Add preference for auto-delete and i18nize strings
Add a preference to prevent auto-deletion of podcasts when they finish
playing or the user skips to the end of the podcast. If the podcasts are
not deleted, then they are moved to the end of the queue - the user is
then responsible for deleting the podcasts, eventually.

While we're in the preferences file, move all of the strings into
strings.xml to support translation. Also, add categories to the
preferences to differentiate the podcast lifecycle management vs.
hardware controls vs. crash reporting & usage data.

Signed-off-by: Dan Scott <dan@coffeecode.net>